+++ OpenMarket/OpenMarket/Service/ProductService.swift
@@ -1,27 +1,6 @@
 import Foundation
 
-struct ProductService {
-    private func doDataTask(
-        with request: URLRequest,
-        session: URLSessionProtocol,
-        completionHandler: @escaping (Result<Data, NetworkingError>) -> Void
-    ) {
-        let task = session.dataTask(with: request) { data, response, error in
-            guard error == nil else {
-                return completionHandler(.failure(.request))
-            }
-            guard let response = response as? HTTPURLResponse,
-                  (200...299).contains(response.statusCode) else {
-                return completionHandler(.failure(.response))
-            }
-            guard let data = data else {
-                return completionHandler(.failure(.data))
-            }
-            completionHandler(.success(data))
-        }
-        task.resume()
-    }
-
+struct ProductService: APIService {
     func retrieveProduct(
         productIdentification: Int,
         session: URLSessionProtocol,
